requirements.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.6.0